// src/app/admin/assessments/assessments.enums.ts
// Define Assessment Module Enums below

export enum AssessmentsModuleEnum {
  SavingAssessmentStatus = 'Saving Assessment...',
  SavedAssessmentStatus = 'Assessment Saved!',
  FailedAssessmentStatus = 'Failed To Save Assessment',
  DraftAssessmentStatus = 'Draft',
  PublishedAssessmentStatus = 'Published',
  ArchivedAssessmentStatus = 'Archived',
  DefaultAssessmentLevel = 'Beginner',
  DefaultTestLevel = 'Coding',
  AllAssessmentStatus = 'All',
  YetToStartAssessmentStatus = 'YetToStart',
  InProgressAssessmentStatus = 'InProgress',
  CompletedAssessmentStatus = 'Completed',
  AssessmentLevel = 'AssessmentLevel',
  TaskType = 'TaskType',
  PackageStatus = 'PackageStatus'
}